Clean up RTC and move it out of Gayle.c
